requirements.txt
Django==4.1
mysqlclient
Pillow
django-filer
django-admin-sortable2